pyproject.toml
[project]
name = "lastfm-export"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "hatchling>=1.28.0",
    "pytest>=9.0.2",
    "responses>=0.25.8",
    "ruff>=0.15.1",
]

[project.scripts]
lastfm-export = "lastfm_export.cli.app:main"
